Moving to Observation instead of List
