+++ data/Templates/eCR/ValueSet/ValueSet.json
@@ -851,20 +851,6 @@
       "__default__": {
         "code": "unknown"
       }
-    },
-    "ValueSet/ListStatus": {
-      "active": {
-        "code": "current"
-      },
-      "inactive": {
-        "code": "retired"
-      },
-      "error": {
-        "code": "entered-in-error"
-      },
-      "__default__": {
-        "code": "unknown"
-      }
     }
   }
 }
